+++ src/adaf_attack/capabilities/dmsa_ops.py
@@ -8,8 +8,8 @@ from typing import Any
 from ldap3 import MODIFY_REPLACE
 from rich.console import Console
 
+from adaf_attack.capabilities.capability_catalog import register_from_catalog
 from adaf_attack.capabilities.gmsa_laps_enum import _parse_managed_password_blob
-from adaf_attack.capabilities.planned_offensive import register_from_catalog
 from adaf_attack.core.graph import AttackGraph
 from adaf_attack.core.ldap_ops import (
     attr_value,