+++ src/adaf_attack/capabilities/maq_ops.py
@@ -7,7 +7,7 @@ from typing import Any
 
 from rich.console import Console
 
-from adaf_attack.capabilities.planned_offensive import register_from_catalog
+from adaf_attack.capabilities.capability_catalog import register_from_catalog
 from adaf_attack.capabilities.rbcd import Rbcd
 from adaf_attack.capabilities.s4u_abuse import S4uAbuse
 from adaf_attack.core.graph import AttackGraph