+++ src/adaf_attack/capabilities/relay_ops.py
@@ -8,7 +8,7 @@ from typing import Any
 
 from rich.console import Console
 
-from adaf_attack.capabilities.planned_offensive import register_from_catalog
+from adaf_attack.capabilities.capability_catalog import register_from_catalog
 from adaf_attack.core.graph import AttackGraph
 from adaf_attack.core.ldap_ops import (
     attr_value,